Cross-Validated Posterior Classification Summaries for a Naive Bayes model

Given a set of observed data including a categorical response variable y and a naiveBayes model of y, this function returns a cross validated confusion matrix by which to assess the model's posterior classification quality.

naive_classification_summary_cv(model, data, y, k = 10)

Arguments

model
a naiveBayes model object with categorical y

data
data frame including the variables in the model

y
a character string indicating the y variable in data

k
the number of folds to use for cross validation

Value

a list

Examples

data(penguins_bayes, package = "bayesrules")
example_model <- e1071::naiveBayes(species ~ bill_length_mm, data = penguins_bayes)
naive_classification_summary_cv(model = example_model, data = penguins_bayes, y = "species", k = 2)

#> $folds
#>   fold    Adelie Chinstrap    Gentoo overall_accuracy
#> 1    1 0.9610390 0.3928571 0.7611940        0.7906977
#> 2    2 0.9733333 0.0500000 0.8947368        0.7325581
#> 
#> $cv
#>    species       Adelie   Chinstrap       Gentoo
#>     Adelie 96.71% (147)  0.00%  (0)  3.29%   (5)
#>  Chinstrap  5.88%   (4) 19.12% (13) 75.00%  (51)
#>     Gentoo  6.45%   (8) 11.29% (14) 82.26% (102)
#>