Cards Against Humanity's Pulse of the Nation Survey

Cards Against Humanity's "Pulse of the Nation" project (https://thepulseofthenation.com/) conducted monthly polls into people's social and political views, as well as some silly things. This data includes responses to a subset of questions included in the poll conducted in September 2017.

pulse_of_the_nation

Format

A data frame with observations on 1000 survey respondents with 15 variables:

income

income in \$1000s

age

age in years

party

political party affiliation

trump_approval

approval level of Donald Trump's job performance

education

maximum education level completed

robots

opinion of how likely their job is to be replaced by robots within 10 years

climate_change

belief in climate change

transformers

the number of Transformers film the respondent has seen

science_is_honest

opinion of whether scientists are generally honest and serve the public good

vaccines_are_safe

opinion of whether vaccines are safe and protect children from disease

books

number of books read in the past year

ghosts

whether or not they believe in ghosts

fed_sci_budget

respondent's estimate of the percentage of the federal budget that is spent on scientific research

earth_sun

belief about whether the earth is always farther away from the sun in winter than in summer (TRUE or FALSE)

wise_unwise

whether the respondent would rather be wise but unhappy, or unwise but happy

Source

https://thepulseofthenation.com/downloads/201709-CAH_PulseOfTheNation_Raw.csv